Apiv3 playoff advancement (#2510)
* Refactor playoff advancement code into own file

* [apiv3] Add event playoff advancement endpoint

Closes #2509
