+++ templates_jinja2/playoff_partials/playoff_round_robin_6_team.html
@@ -8,15 +8,9 @@
         <th>Teams</th>
         <th>Record</th>
         <th>Champ Points</th>
-        {% if event.year == 2018 %}
-          <th>Park/Climb Points</th>
-          <th>Auto Points</th>
-        {% elif event.year == 2019 %}
-          <th>Cargo Points</th>
-          <th>Hatch Panel Points</th>
-        {% else %}
-          <th>Match Points</th>
-        {% endif %}
+        {% for tiebreaker in playoff_advancement_tiebreakers %}
+          <th>{{tiebreaker}}</th>
+        {% endfor %}
         <th>Advance to Finals</th>
       </tr>
     </thead>
@@ -29,7 +23,7 @@
           <td>{{ record.wins }}-{{ record.losses }}-{{ record.ties }}</td>
           <td>{{ champ_points }}</td>
           <td>{{ tiebreaker1 }}</td>
-          {% if event.year >= 2018 %}
+          {% if playoff_advancement_tiebreakers|length == 2 %}
             <td>{{ tiebreaker2 }}</td>
           {% endif %}
           <td>{% if playoff_advancement.sf_complete and loop.index0 < 2 %}<span class="glyphicon glyphicon-ok" aria-hidden="true"></span>{% endif %}</td>
